// src/types/quoteSummary.types.ts
export interface QuoteSummaryData {
  number_of_adults: number;
  number_of_children: number;
  number_of_infants: number;
  accommodation_cost: number;
  transport_cost: number;
  transfer_cost: number;
  excursion_cost: number;
  net_cost: number;
  markup_cost: number;
  total_cost: number;
  markup_percentage: number;
}